+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+ 
+ - add new role moderator and permission to manage moderators
+
 ## [1.0.0-beta.5] - 2020-03-01
 
 ### Changed